requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
pymysql
python-dotenv
apscheduler 
joblib
firebase-admin